requirements.txt
discord.py
PyYAML
requests
